requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-appconf==1.1.0
django-geoip==0.5.2.1
django-ipware==7.0.1
django-ratelimit==4.1.0
django-redis==6.0.0
frozenlist==1.7.0
geoip2==5.1.0
idna==3.10
maxminddb==2.8.2
multidict==6.6.4
propcache==0.3.2
python-ipware==3.0.0
redis==6.4.0
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
